Adjust to compile cleanly with MSVC
Add CI job which compiles and runs C tests which uses `-W3 -WX` (`-W3`
is apparently the default warning level when using the IDE and `-WX`
turns warnings into errors).

The build system isn't really set up properly for this currently - it
has `.o` hard-wired as the object extension (which MSVC handles but
moans about) and builds a static library with extension `.a` using the
mingw `ar` tool rather than the MSVC tooling.

The build does at least seem to work correctly, and the key thing we
want to ensure here is that the generated stemmer code and runtime
support works, which our CI now does.

+++ compiler/analyser.c
@@ -881,7 +881,7 @@ static struct node * make_among(struct analyser * a, struct node * p, struct nod
 
         for (w = w0 - 1; w >= v; w--) {
             if (w->size < size && memcmp(w->b, b, w->size * sizeof(symbol)) == 0) {
-                w0->i = w - v;  /* fill in index of longest substring */
+                w0->i = (int)(w - v);  /* fill in index of longest substring */
                 break;
             }
         }
